add ffAppUpdatePrompt feature flag and update featureControls imports
