+++ libs/expo/betterangels/src/lib/providers/featureControls/types.ts
@@ -1,4 +1,8 @@
-// Define a dictionary type for easier access
+import { FeatureFlags } from './constants';
+
+export type TFeatureFlagKey = keyof typeof FeatureFlags;
+export type TFeatureFlagValue = (typeof FeatureFlags)[TFeatureFlagKey];
+
 export interface FeatureControlDictionary {
   [key: string]: { isActive: boolean; lastModified?: string | null };
 }
