+++ libs/expo/betterangels/src/lib/providers/featureControls/useFeatureControls.ts
@@ -1,0 +1,17 @@
+import { useContext } from 'react';
+import {
+  FeatureControlContext,
+  TFeatureControlContext,
+} from './featureControlContext';
+
+export const useFeatureControls = (): TFeatureControlContext => {
+  const context = useContext(FeatureControlContext);
+
+  if (context === undefined) {
+    throw new Error(
+      'useFeatureControls must be used within a FeatureControlProvider'
+    );
+  }
+
+  return context;
+};
