+++ libs/expo/betterangels/src/lib/providers/index.ts
@@ -1,5 +1,4 @@
-export { useFeatureControls } from './featureControls/featureControlContext';
-export { FeatureControlProvider } from './featureControls/featureControlProvider';
+export * from './featureControls';
 export { default as KeyboardToolbarProvider } from './keyboardToolbar/keyboardToolbarProvider';
 export { default as SnackbarProvider } from './snackbar/SnackbarProvider';
 export { default as UserProvider } from './user/UserProvider';
