+++ libs/expo/betterangels/src/lib/screens/UserProfile/index.tsx
@@ -7,10 +7,9 @@ import {
   TextBold,
 } from '@monorepo/expo/shared/ui-components';
 import { router } from 'expo-router';
-import { StyleSheet, Text, View } from 'react-native';
+import { StyleSheet, View } from 'react-native';
 import { useDeleteCurrentUserMutation } from '../../apollo';
 import { useSignOut, useSnackbar, useUser } from '../../hooks';
-import { useFeatureControls } from '../../providers/featureControls/featureControlContext';
 import InfoCard from './InfoCard';
 
 export default function UserProfile() {
@@ -18,7 +17,6 @@ export default function UserProfile() {
   const { showSnackbar } = useSnackbar();
 
   if (!user) throw new Error('Something went wrong');
-  const featureControls = useFeatureControls();
   const [deleteCurrentUser] = useDeleteCurrentUserMutation();
   const userInfo = [
     { title: 'Email', value: user.email },
@@ -46,9 +44,6 @@ export default function UserProfile() {
       });
     }
   }
-
-  const isSamplFeatureFlagActive =
-    featureControls.flags['SampleFeatureFlag']?.isActive;
 
   return (
     <View style={styles.container}>
@@ -90,11 +85,6 @@ export default function UserProfile() {
           }
         />
       </View>
-      {isSamplFeatureFlagActive && (
-        <Text style={styles.featureText}>
-          The feature "SampleFeatureFlag" is active
-        </Text>
-      )}
     </View>
   );
 }
